+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add git revision to "sr version" output
 - Rename `remove-reviewed` embedded step to `skip-reviewed` for clarity. `remove-reviewed` still works as an alias.
+- Rename `review` command to `flow`. `review` still works as an alias.
 
 ## v0.12.0 (2023-01-10)